Enum clarith::Number

pub enum Number {
    Special(Special),
    Other(Option<Primer>, Clog),
}

A finite unbounded number with unbounded precision.

Variants
Special(Special)Other(Option<Primer>, Clog)
Methods

impl Number

pub fn unwrap_special(self) -> Special

Unwraps the Special content from self, or panic.

pub fn unwrap_other(self) -> (Option<Primer>, Clog)

Unwraps the Other content from self, or panic.

pub fn compare(n1: Number, n2: Number) -> Ordering

Destructively compare two Numbers.

pub fn ratio(num: isize, den: isize) -> Number

Construct a Number from the ratio of two signed machine integers.

pub fn ratio_u(positive: bool, num: usize, den: usize) -> Number

Construct a Number from the ratio of two unsigned machine integers.

pub fn homographic(
x: Number,
nx: isize,
n: isize,
dx: isize,
d: isize
) -> Number

Construct the Number (nx * x + n) / (dx * x + d)

pub fn combine(
x: Number,
y: Number,
nxy: isize,
nx: isize,
ny: isize,
n: isize,
dxy: isize,
dx: isize,
dy: isize,
d: isize
) -> Number

Construct the Number (nxy * x * y + nx * x + ny * y + n) / (dxy * x * y + dx * x + dy * y + d)

pub fn consume(x: Number) -> (Option<Number>, isize, isize, isize, isize)

Transfer information from a Number to a homograhic transformation.

This funtion returns:

A Number y

Coefficients ny, n, dy, d

such that (ny * y + n) / (dy * y + d) equals to the input x.

The information extraction from the input x proceeds until one of the following conditions is met:

The input reaches one half. (In this case y is returned None.)

The returned coefficients are about to overflow.